feat: Add gossip and eviction to the mempool (#200)
This PR adds a basic implementation of the gossipped mempool described
in #180.

It uses the existing mempool implementation but changes the key to use a
key derived from the timestamp and identity of a message. Additions to
the mempool are broadcast using a new gossip message. Messages are also
evicted from the mempool when shard chunks are emitted that contain
messages contained in the mempool.

Some minor changes to the channels and overall plumbing was required to
setup the mempool communication.