only check for existence in mempool on initial insert (#275)
There's no reason to perform this check when we pull messages out of the
mempool because the message was just popped and will never exist in the
mempool. This check scales with `log(size of mempool)` and it becomes
more expensive as the mempool gets larger so we should avoid doing it if
possible.