perf: Tune consensus timeouts and dynamically adjust commit delay for consistent block times (#308)
It takes us ~250-300ms to propose and validate a block with 1K messages
under load. Propose timeout has to include sufficient time for both the
making the proposal and another node to receive and validate it. So, use
a 1s propose timeout.

Also remove the commit delay which always waits 250ms. Instead
dynamically target a 1s block time based on how long it took to decide a
block. Do not wait if we're behind schedule