perf: Cache transaction on propose and validate to replay on commit (#309)
Each node replays the transactions twice, once either in propose or
validate and then once more in decide. Caching and replaying on decide
will save about ~30% of the time spent by the node to decide a height.
With 1K msgs/block it takes:

~300ms to propose + 300 ms to validate + 300 ms to validate again +
~25ms to commit

This change will bring it down to ~625ms from ~925ms.