[feat] Add new event type for merge failures (#374)
Add a new HubEvent type for merge failures that contains the error code
and reason why message failed to be merged. This is useful for clients
who submit messages which were included into the mempool, but then
failed later.

Also updates some api error messages to be consistent with hubs, as well
as return some additional fields in `GetInfo` and `HubEvent`