fix: allow http urls to be missing query parameters for optional fields (#408)
We really want `default` wherever we have `skip_serializing_if` for the
request types. We never serialize these types--we only deserialize them
and we want to default to `None` if these fields are missing.

Right now we error if `pageToken` is not specified but we want such a
request to go through.