fix: `getidregistryonchaineventbyaddress` (#471)
Implements `get_id_registry_on_chain_event_by_address`

I fought with how to handle the `get_onchain_events` we could
alternatively make the `fid` on that optional but I didn't want to mess
with the existing interfaces too much. I also initially added an
iterator on the events but backtracked on that since it didn't seem
right but you can see if in the previous commits. I ended up adding a
`get_all_onchain_events` which doesn't take an FID at all. If we'd
prefer to make the `fid` optional on the existing function I can do
that.

I added a test and it's passing but we might want to put the iterator
through the ringer a little more. Also callout I did not test this with
complete production state so I don't know how much this will effect
latency on a cold get before the cache spins up so that might be worth
testing but I don't have the full snapchain local right now (waiting on
a drive to get that running again)

@aditiharini 

Closes #470