fix: Mempool should not do duplicate checks for validator messages (#545)
The mempool duplicate message check is a performance optimization.
Onchain events and fname transfers are not that frequent, and don't
benefit from this. The engine is capable of handling duplicates.


This check was prevent fname transfers from being recorded correctly,
because it was checking if the name proof existed at all, and was not
specific to the fid.