feat: Make snapshot metadata compatible with snapdown (#597)
Fixes https://github.com/farcasterxyz/snapchain/issues/560

This fetches all metadata in one go and makes it compatible with
snapdown so that it's possible to use snapdown to resume broken
downloads.

Also add a check to make sure the downloaded file size matches the
content length similar to snapdown so we catch and retry file corruption
early.