feat: Add SubmitBulkMessages rpc/http API (#600)
Add the SubmitBulkMessages API (which had been commented out). 

The endpoint (both grpc and http API) accepts an array of messages, and
submits them sequentially. Right now, the messages need to be
independent (i.e., a message that depends on a previous message in the
array to succeed is not supported yet)

It will attempt to merge all the messages, and return success/failure
for each one individually